+++ src/metainfo.rs
@@ -45,7 +45,7 @@ pub struct Info {
     pub piece_length: u32,
     pub pieces: Vec<Sha1>,
     pub name: String,
-    pub length: u32,
+    pub length: u64,
 }
 
 impl FromBencode for Info {
